QcomModulePkg: Correct the image header size
Block size is not equal to page size. Image header check will be failed
when the block size is less than the struct size of boot image header.
So use the max boot image page size for image header size. As it's hard
to get the page size before loading boot image.

Change-Id: I492e3bc60a1fd295a4503a67dbf1f7fb2f0b828a